% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arraytogroundandcanopyrasters.R
\name{array.to.ground.and.canopy.rasters}
\alias{array.to.ground.and.canopy.rasters}
\title{Convert a list of voxelized arrays into two rasters - ground and canopy height}
\usage{
array.to.ground.and.canopy.rasters(laz.array, epsg.code)
}
\arguments{
\item{laz.array}{Original LiDAR pulse count array that was generated using the laz.to.array function. This contains
spatial information for the array.}

\item{epsg.code}{EPSG code so that the rasters can be projected into the appropriate projection}
}
\value{
A list of a ground and canopy height raster.
}
\description{
This function reads in a list of voxelized arrays, converts each one to ground and canopy height rasters, and then merges them all together
to form a ground and canopy height raster of the entire study area.
}
