% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopyheightlevelr.R
\name{canopy.height.levelr}
\alias{canopy.height.levelr}
\title{Level a voxelized array to the ground, mimicing a canopy height model}
\usage{
canopy.height.levelr(lidar.array.list)
}
\arguments{
\item{lidar.array.list}{List of voxelized arrays that was created with the las.to.array function}
}
\value{
A list of leveled voxelized arrays
}
\description{
This function reads in a list of voxelized arrays and levels them to the ground, thus each array
will mimic a canopy height model, making comparisions between voxels easier
}
