% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopyvolume.R
\name{canopy.volume}
\alias{canopy.volume}
\title{Convert LAD estimates into fives rasters explaining the volume and total leaf area of the euphotic (portion of the
canopy where 65% of the leaf material is located) and the oligophotic zone (the remaining canopy) and the volume
of empty space within the canopy (from top of the canopy the ground)}
\usage{
canopy.volume(lad.array, laz.array, ht.cut, xy.res, z.res, epsg.code)
}
\arguments{
\item{lad.array}{LAD estimate array that was generated using the machorn.lad function.}

\item{laz.array}{Voxelized LiDAR array that was generated using the laz.to.array function. This contains
spatial information for all arrays.}

\item{ht.cut}{Height that calculations will exclude. This is to remove understory LAD estimates from
further calculations. If 5 is entered then all voxels 5 meters and above will be included. Enter 0 if
you want to include all calculations}

\item{xy.res}{Horizontal resolution of each voxel - if it is 10x10 meters then just enter 10}

\item{z.res}{Vertical resolution of each voxel - if it is 1 meter then just enter 1}

\item{epsg.code}{EPSG code so that the rasters can be projected into the appropriate projection}
}
\value{
A list containing the quantile and mean rasters.
}
\description{
This function reads in a the LAD estimates that were previously calculated,
calculates the height cutoff for the euphotic zone, calculates the volume and total leaf area of this
portion of the canopy and then does the same for the remaining canopy. It then returns 5 rasters
}
\details{
This function also uses a 3x3 moving window to calculate the volume of the euphotic, oligophotic, and empty
zones within the moving window. This returns 5 additional rasters, one for each variable.

These forest structure attributes are based off calculations from:

Lefsky, M.A., Cohen, W.B., Acker, S.A., Parker, G.G., Spies, T.A., and Harding, D. (1999).
Lidar Remote Sensing of the Canopy Structure and Biophysical Properties of Douglas-Fir Western Hemlock Forests.
Remote Sensing of the Environment, 70, 339-361. https://doi.org/10.1016/S0034-4257(99)00052-8
}
