% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladarraytorasterstack.R
\name{lad.array.to.raster.stack}
\alias{lad.array.to.raster.stack}
\title{Convert a list of voxelized arrays into a single raster stack}
\usage{
lad.array.to.raster.stack(lad.array.list, epsg.code)
}
\arguments{
\item{lad.array.list}{List of voxelized LAD arrays that was created with the machorn.lad function}

\item{epsg.code}{EPSG code so that the rasters can be projected into the appropriate projection}
}
\value{
A raster stack of LAD estimates. Each raster is associated with a slice of the canopy. The Z resolution was determined
in the laz.to.array function
}
\description{
This function reads in a list of voxelized arrays, converts each one to a raster stack, and then merges them all together
to form a raster stack of the entire study area. This function also creats a lot of extra NA layers in the raster stack so that
the individual rasters can easily be merged together.
}
