% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladquantiles.R
\name{lad.quantiles}
\alias{lad.quantiles}
\title{Convert LAD estimates into six rasters explaining the height distribution of LAD within the canopy
 - the heigth of the 10th, 25th, 50th, 75th, 90th quantiles aas well as the mean}
\usage{
lad.quantiles(lad.array, laz.array, ht.cut, epsg.code)
}
\arguments{
\item{lad.array}{LAD estimate array that was generated using the machorn.lad function.}

\item{laz.array}{Voxelized LiDAR array that was generated using the laz.to.array function. This contains
spatial information for all arrays.}

\item{ht.cut}{Height that calculations will exclude. This is to remove understory LAD estimates from
further calculations. If 5 is entered then all voxels 5 meters and above will be included. Enter 0 if
you want to include all calculations}

\item{epsg.code}{EPSG code so that the rasters can be projected into the appropriate projection}
}
\value{
A list containing the quantile and mean rasters.
}
\description{
This function reads in a the LAD estimates that were previously calculated, 
calculates the height where a host of different quantiles and the mean occcur and return a raster
showing the height where each quantile occurs within a given vertical column.
}
\details{
These forest structure attributes are based off calculations from:

Shi, Y., Wang, T., Skidmore, A.K., and Heurich, M. (2018). Important LiDAR metrics for discriminating 
forest tree species in Central Europe. ISPRS Journal of Photogrammetry and Remote Sensing, 137, 
163-174. https://doi.org/10.1016/j.isprsjprs.2018.02.002
}
