% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laztoarray.R
\name{laz.to.array}
\alias{laz.to.array}
\title{Convert a .laz or .las file to an array}
\usage{
laz.to.array(laz.files.path, voxel.resolution, z.resolution,
  use.classified.returns)
}
\arguments{
\item{laz.files.path}{Path to the folder containing the .laz or .las files}

\item{voxel.resolution}{The spatial resolution (x,y) you want the output voxel to be - this is a single
number where both sides of the cell will be the same}

\item{z.resolution}{the vertical resolution of the voxel}

\item{use.classified.returns}{TRUE or FALSE statement to determine if the user wants to use preclassified ground returns or the lowest point for "ground"}
}
\value{
A list of voxelized arrays
}
\description{
This function reads in a list of .laz or .las files from a given file path and converts
each one into a voxelized array based on the x,y,z resolutions you specify.
}
