% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machornlad.R
\name{machorn.lad}
\alias{machorn.lad}
\title{Convert a lidar array into an array of LAD estimates}
\usage{
machorn.lad(leveld.lidar.array, voxel.height,
  beer.lambert.constant = NULL)
}
\arguments{
\item{leveld.lidar.array}{Leveled lidar array, which is the output of the canopy.height.levelr function}

\item{voxel.height}{The Z parameter that was entered into the laz.to.array function}

\item{beer.lambert.costant}{If you do not have a Beer Lambert Constant then set this NULL. If you have one, then enter it here.}
}
\value{
A list of voxelized array which contain LAD estimates within each voxel
}
\description{
This function reads in the lidar array created with the laz.to.array function and uses the MacArthur and Horn method to
estimate LAD within each voxel. The sum of each column is equal to the LAI of that column of voxels.
}
