% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocrugosity.R
\name{toc.rugosity}
\alias{toc.rugosity}
\title{Convert canopy height model into a top of canopy rugosity raster}
\usage{
toc.rugosity(chm.raster, xy.res, z.res)
}
\arguments{
\item{chm.raster}{CHM that was generated using the array.to.ground.and.canopy.rasters function.}

\item{xy.res}{Horizontal resolution of the CHM raster - if it is 10x10 meters then enter 10}

\item{z.res}{vertical resolution of the original LAD estimates - if it is 1 meter then enter 1}
}
\value{
A raster showing the volume of empty space at the top of the canopy.
}
\description{
This function reads in a the CHM previously calculated and determines the amount of empty space at the top
of the canopy compared to the other trees within a 3x3 moving window. The output is the volume of this area.
}
\details{
These forest structure attributes are based off calculations from:

Lefsky, M.A., Cohen, W.B., Acker, S.A., Parker, G.G., Spies, T.A., and Harding, D. (1999).
Lidar Remote Sensing of the Canopy Structure and Biophysical Properties of Douglas-Fir Western Hemlock Forests.
Remote Sensing of the Environment, 70, 339-361. https://doi.org/10.1016/S0034-4257(99)00052-8
}
