% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withincanopyrugosity.R
\name{within.canopy.rugosity}
\alias{within.canopy.rugosity}
\title{Convert LAD estimates into two rasters - standard deviation of LAD in a given column and canopy rugosity}
\usage{
\method{within}{canopy.rugosity}(lad.array, laz.array, ht.cut, epsg.code)
}
\arguments{
\item{lad.array}{LAD estimate array that was generated using the machorn.lad function.}

\item{laz.array}{Voxelized LiDAR array that was generated using the laz.to.array function. This contains
spatial information for all arrays.}

\item{ht.cut}{Height that calculations will exclude. This is to remove understory LAD estimates from
further calculations. If 5 is entered then all voxels 5 meters and above will be included. Enter 0 if
you want to include all calculations}

\item{epsg.code}{EPSG code so that the rasters can be projected into the appropriate projection}

\item{xy.res}{Resolution of xy coordinates - if it is a 10x10 meter pixel then enter 10 here}

\item{z.res}{Vertical resolution of voxel - if it is 1 meter tall then enter 1}
}
\value{
A list containing max LAD and height of max LAD rasters.
}
\description{
This function reads in a the LAD estimates that were previously calculated, 
finds the standard deviation of LAD estimates within a given column, then uses a 3x3 moving window
to find the standard deviation of the previously calculated vertical standard deviations in the moving window.
This value is then saved in a new matrix, so that the value calculated by the moving window is not used in
subsequent calcluations. This is the Hardiman approach to calcluating canopy rugosity to look at within
canopy variation of LAD.
}
\details{
Because the second half of this funtion uses a 3x3 moving window, the outside rows and columns are lost due to
insufficient data. If you are mosaicing multiple rasters together in later steps, this needs to be taken into 
account. I am working on a function that will pass a moving window over a large raster, but it is still in 
development.

These forest structure attributes are based off calculations from:

Hardiman, B. S., Bohrer, G., Gough, C. M., Vogel, C. S., & Curtis, P. S. (2011). 
The role of canopy structural complexity in wood net primary production of a maturing northern deciduous forest. 
Ecology, 92, 1818-1827. https://doi.org/10.1890/10-2192.1
}
